Correction des warning visual de Windows

+++ include/Polynomials/APolynomial.hpp
@@ -38,7 +38,7 @@ public:
                       VectorVectorDouble& out) const {};
   VectorDouble evalOp(cs* Op, const VectorDouble& in) const;
   VectorDouble getCoeffs()const{ return _coeffs;}
-  int getDegree() const { return _coeffs.size();}
+  int getDegree() const { return static_cast<int>(_coeffs.size());}
   virtual int fit(std::function<double(double)> f,
                   double from = 0.,
                   double to = 1.,
